% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{annotateAFG}
\alias{annotateAFG}
\title{Calculate labels (m/z differences) for vertices and edges of a pattern}
\usage{
annotateAFG(fp, atoms, dags, elabs, oformula, edge_label = "lab")
}
\arguments{
\item{fp}{id of the pattern}

\item{atoms}{vector of atom names to consider (for example, CHNOPS)}

\item{dags}{vector of the fragmentation graphs (DAGs) contained in the pattern}

\item{elabs}{dataframe of all the discretized edge labels (m/z differences) in the dataset}

\item{oformula}{vector of precursor formula for each spectra in the pattern, if given}

\item{edge_label}{id of the field containing the m/z differences in elabs}
}
\description{
For a specific pattern, this function calculates the labels (m/z differences) for vertices and edges.
Labels for the vertices correspond to m/z differences from the root of the pattern (precursor parent).
}
\details{
This function should not be used by the user.
}
