% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling_gnps.R
\name{annotateNetwork}
\alias{annotateNetwork}
\title{Title}
\usage{
annotateNetwork(components, net, patterns, copy = TRUE,
  keepattr = c("Compound_Name", "SpectrumID", "DefaultGroups",
  "cluster index", "precursor mass"), sep_infos = ",")
}
\arguments{
\item{components}{jkkj}

\item{net}{kjkj}

\item{patterns}{jkkj}

\item{copy}{kjkj}

\item{sep_infos}{kjkj}
}
\description{
Title
}
\examples{
print("Examples to be put here")
}
