% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFormula.R
\docType{methods}
\name{as.character,LossFormula-method}
\alias{as.character,LossFormula-method}
\title{Return a string prerpesentation of a lossFormula object}
\usage{
\S4method{as.character}{LossFormula}(x)
}
\arguments{
\item{x}{A lossFormula object.}
}
\value{
A character vector with all the possible formula for a loss object.
}
\description{
Return a string presentation of all the possible formula. Should not be used by the user.
}
