% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern.R
\name{calculateCoverage,fragPattern-method}
\alias{calculateCoverage,fragPattern-method}
\title{Calculate the coverage of a specific pattern}
\usage{
\S4method{calculateCoverage}{fragPattern}(x, mzloss, mgs)
}
\arguments{
\item{x}{The pattern}

\item{mzloss}{The table of m/z differences}

\item{mgs}{the m/z graphs (the DAGs)}
}
\value{
The m2l object with all the coverage calculated.
}
\description{
This function should never be called by the user. Call the ms2Lib method
}
