% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/fragPattern.R, R/ms2Lib.R
\name{calculateCoverage}
\alias{calculateCoverage}
\alias{calculateCoverage,fragPattern-method}
\alias{calculateCoverage,ms2Lib-method}
\title{Calculate the relative intensity covered by patterns.}
\usage{
calculateCoverage(x, ...)

\S4method{calculateCoverage}{fragPattern}(x, mzloss, mgs)

\S4method{calculateCoverage}{ms2Lib}(x)
}
\arguments{
\item{x}{The ms2Lib to be computed.}

\item{...}{supplementary argument passed to the methods}

\item{mzloss}{The table of m/z differences}

\item{mgs}{the m/z graphs (the DAGs)}
}
\value{
The m2l object with all the coverage calculated.
}
\description{
`calculateCoverage` calculates the relative intensity covered by a pattern or a library of patterns.

The S4 method for the fragPattern object should never be called by the user. Call the ms2Lib method instead.
}
\examples{
#Loading the data
data(m2l)

#Calculate the coverage for an ms2Lib object
m2l <- calculateCoverage(m2l)
}
