% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\name{createFileCS}
\alias{createFileCS}
\title{Create a png file containing the 2D structure of the given molecule
(obtained from ChemSpider with the inchi key of the molecule, needs an API key)}
\usage{
createFileCS(name, path_inchi, dir_images)
}
\arguments{
\item{name}{the name of the molecule in the tsv file given by path_inchi}

\item{path_inchi}{name of a tabular file containing the inchi keys of the molecules in a column named "name"; if this table is available, the 2D structures corresponding to the spectra will be retrieved from ChemSpider (webchem package) and displayed in the plot along the spectra (default NULL)}

\item{dir_images}{the path to the directory to store the png images}
}
\value{
True if the molecule is found, False otherwise
}
\description{
Create a png file containing the 2D structure of the given molecule
(obtained from ChemSpider with the inchi key of the molecule, needs an API key)
}
