% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeLabels.R
\docType{methods}
\name{discretizeMassFragments,ms2Lib-method}
\alias{discretizeMassFragments,ms2Lib-method}
\title{Discretize the fragments of mass spectrometers.}
\usage{
\S4method{discretizeMassFragments}{ms2Lib}(m2l, adducts = NULL,
  ppm = 7, dmz = 0.002, mode = c("POS", "NEG"), count = 2,
  limMzFormula = c(50, 250), maxFrags = 15, maxOverlap = 0.05,
  strictMatching = TRUE, precPpm = 20, precDmz = 0.02,
  atoms = NULL, heteroAtoms = TRUE, mzDigits = 4, ...)
}
\arguments{
\item{m2l}{The ms2 lib object to be discretized.}

\item{adducts}{A ppm vector giving the adducts names. ADD FUNCTION.}

\item{ppm}{the maximum authorized deviation in ppm (parts per million).}

\item{dmz}{The maximum authorized deviation in Da.}

\item{count}{The minimum number of spectra in which the label needs to be found. Shall be greater or equal to 2.}

\item{limMzFormula}{An interval giving the range in which the formula will be calculated.
peaks with a masses lower than the lower term will be ignored while peak with a mass higher than the
higher term won't have formula check.}

\item{maxFrags}{The maximum number of fragment allower on the spectra.}

\item{maxOverlap}{The degree of overlap allowed between bins. Bins which overlap more are fused.}

\item{strictMatching}{Shall the formula matching be strict, or approximated.}

\item{precPpm}{The ppm tolerance used ot match the precursor mz to the fragments.}

\item{precDmz}{The minimum tolerance used to match the precursor mz to the fragments.}

\item{atoms}{A list of the maximum number of atoms authorized during the formula generation process.
If it is NULL, the default list is C:limMzFormula[2]%/%12, H:50, N:6 ,O:6.}

\item{heteroAtoms}{a boolean,  ignored if a custom atoms is furnished, add heteroatoms P, Cl, S as possible atoms.}

\item{mzDigits}{The number of decimals to write as an information in the graphs.}
}
\value{
An ms2Lib object with filled fileds and constructed dags.
}
\description{
Discretize the fragments in a ms2Lib object.
}
\examples{
print("Examples ot be put here !")
}
