% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeLabels.R
\docType{methods}
\name{discretizeMassLosses,ms2Lib-method}
\alias{discretizeMassLosses,ms2Lib-method}
\title{Discretize a set of mass differences.}
\usage{
\S4method{discretizeMassLosses}{ms2Lib}(m2l, ppm = 7, dmz = 0.002,
  freq = 2/length(m2l), limMzFormula = c(14.5, 200), maxFrags = 15,
  maxOverlap = 0.05, strictMatching = TRUE, precPpm = 20,
  precDmz = 0.02, atoms = NULL, heteroAtoms = TRUE, mzDigits = 4,
  penalizedLosses = NULL, majoredLosses = character(0), ...)
}
\arguments{
\item{m2l}{The ms2 lib object to be discretized.}

\item{ppm}{the maximum authorized deviation in ppm (parts per million).}

\item{dmz}{The maximum authorized deviation in Da.}

\item{freq}{The frequency threshold of the dataset.}

\item{limMzFormula}{An interval giving the range in which the formula will be calculated.
peaks with a masses lower than the lower term will be ignored while peak with a mass higher than the
higher term won't have formula check.}

\item{maxOverlap}{The degree of overlap allowed between bins. Bins which overlap more are fused.}

\item{strictMatching}{Shall the formula matching be strict, or approximated.}

\item{precPpm}{The ppm tolerance used ot match the precursor mz to the fragments.}

\item{precDmz}{The minimum tolerance used to match the precursor mz to the fragments.}

\item{atoms}{A list of the maximum number of atoms authorized during the formula generation process.
If it is NULL, the default list is C:limMzFormula[2]%/%12, H:50, N:6 ,O:6.}

\item{heteroAtoms}{a boolean,  ignored if a custom atoms is furnished, add heteroatoms P, Cl, S as possible atoms.}

\item{mzDigits}{The number of decimals to write as an information in the graphs.}

\item{penalizedLosses}{A chracter vector, the losses which will be penalized. If NULL the table penalizedLossDefault() is used.}

\item{maxFrag}{The maximum number of fragment allower on the spectra.}
}
\value{
An ms2Lib object with filled fileds and constructed dags.
}
\description{
Discretize a set of mass spectra by mass losses and build a set of dags using this.
}
\examples{
print("Examples ot be put here !")
}
