% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{discretizeMassLosses}
\alias{discretizeMassLosses}
\title{Discretize mass differences.}
\usage{
discretizeMassLosses(m2l, ...)
}
\arguments{
\item{m2l}{A ms2Lib object}

\item{...}{Supplmentary arguments passed to the discretization method}
}
\description{
`discretizeMassLosses` discretize the mass differences iun MS-MS spectra.
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
