% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/discretizeLabels.R
\name{discretizeMzDifferences}
\alias{discretizeMzDifferences}
\alias{discretizeMzDifferences,ms2Lib-method}
\title{Discretize m/z differences.}
\usage{
discretizeMzDifferences(m2l, ...)

\S4method{discretizeMzDifferences}{ms2Lib}(
  m2l,
  ppm = 15,
  dmz = 0.007,
  count = 2,
  limMzFormula = c(14.5, 200),
  maxFrags = 15,
  maxOverlap = 0.05,
  strictMatching = TRUE,
  precPpm = 20,
  precDmz = 0.02,
  atoms = NULL,
  heteroAtoms = TRUE,
  mzDigits = 4
)
}
\arguments{
\item{m2l}{The ms2 lib object to be discretized.}

\item{...}{Supplementary arguments passed to the discretization method}

\item{ppm}{the maximum authorized deviation in ppm (parts per million) [default: 15]}

\item{dmz}{The maximum authorized deviation in Da [default: 0.007]}

\item{count}{The minimum number of spectra in which the label needs to be found. Shall be greater or equal to 2 [default: 2]}

\item{limMzFormula}{An interval giving the range in which the formula will be calculated: peaks with a m/z lower than the lower term will be ignored while peak with a mass higher than the higher term won't have formula check [default: c(14.5, 200)]}

\item{maxFrags}{The maximum number of fragment allowed on the spectra [default: 15]}

\item{maxOverlap}{The degree of overlap allowed between bins. Bins with a higher overlap are fused [default: 0.05]}

\item{strictMatching}{Shall the formula matching be strict, or approximated [default: TRUE]}

\item{precPpm}{The ppm tolerance used to match the precursor mz to the fragments [default: 20]}

\item{precDmz}{The minimum tolerance used to match the precursor mz to the fragments [default: 0.02]}

\item{atoms}{A list of the maximum number of atoms authorized during the formula generation process. If set to NULL [default], the default list includes C:max(limMzFormula)/12, H:50, N:10, O:15, Cl:2, S:2, P:2 (if heteroAtoms is set to its default, TRUE), or C:max(limMzFormula)/12, H:50, N:10, O:15 otherwise}

\item{heteroAtoms}{Should heteroatoms P, Cl, S be included as possible atoms (ignored if a custom list of atoms is provided) [default: TRUE]}

\item{mzDigits}{The number of decimals to display on the graphs [default: 4]}
}
\value{
An ms2Lib object with filled fields and constructed dags.
}
\description{
`discretizeMzDifferences` discretizes the m/z differences in MS/MS spectra.

It also builds fragmentation graphs from MS/MS spectra using discretized m/z differences as edges.
}
\examples{
# Loading the data
data(m2l)

# Constructing edge labels
m2l <- discretizeMzDifferences(m2l)

# Setting the maximum number of atoms to specific values
m2l <- discretizeMzDifferences(m2l, 
atoms = list("C"=50,"H"=100,"N"=6,"O"=6,"S"=2,"Cl"=1,"P"=2))
}
