% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling_gnps.R
\name{ExplainComponents}
\alias{ExplainComponents}
\title{Find a pattern for each MS-MS network components.}
\usage{
ExplainComponents(net, m2l, type = c("f1", "recall", "accuracy"),
  vname = "cluster index", ...)
}
\arguments{
\item{net}{An MS-MS network}

\item{m2l}{An ms2Lib object}

\item{type}{The matching algorithm to be used to determine the best pattern}

\item{...}{Supplementaryh parameters to be passed to findAllComponents.}

\item{vnames}{The column used to determine the id.}
}
\description{
At the moment only clique are exported.
}
