% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling_gnps.R
\name{ExplainGNPSComponents}
\alias{ExplainGNPSComponents}
\title{Find a pattern for each MS-MS network components.}
\usage{
ExplainGNPSComponents(net, m2l, metric = c("f1", "recall", "accuracy"),
  vname = "cluster index", threshold = NA_real_)
}
\arguments{
\item{net}{An MS-MS network}

\item{m2l}{An ms2Lib object}

\item{type}{The matching algorithm to be used to determine the best pattern}

\item{vnames}{The column used to determine the id.}

\item{...}{Supplementary parameters to be passed to findGNPSComponents.}
}
\description{
At the moment only clique are exported.
}
\examples{
print("Example to be put here")
}
