% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportLattice.R
\docType{methods}
\name{exportLattice,ms2Lib-method}
\alias{exportLattice,ms2Lib-method}
\title{Export a lattice}
\usage{
\S4method{exportLattice}{ms2Lib}(m2l, filename, format = "graphml", ...)
}
\arguments{
\item{m2l}{A ms2Lib object.}

\item{format}{The format of export in a file.}

\item{file}{A file connection ot a string giving th ename of a file.}
}
\value{
Nothing
}
\description{
Export a lattice correspoding to an ms2Lib object to a file.
}
\examples{
print("Examples to be put here")
}
