% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filterPatterns}
\alias{filterPatterns}
\title{Filter the patterns with intensity coverage}
\usage{
filterPatterns(
  m2l,
  filter_function = function(x, threshold, num) {
     return(sum(x@occurrences[,
    "coverage"] > threshold) >= num)
 },
  args = list(threshold = 0.2, num = 2)
)
}
\arguments{
\item{m2l}{The ms2Lib object ot be filtered}

\item{filter_function}{The filtering function this function take as input a pattern and return TRUE or FALSE.}

\item{args}{A list of supplementary argument to filter_function}
}
\value{
The supplementary ms2Lib arguments (reducedPatterns argument)
}
\description{
A function to filter the patterns according to their coverage of the dataset. 
If less than two spectra in the pattern have a coverage above a threshold of 0.2, then the pattern is discarded.
}
\examples{
m2l <- filterPatterns(m2l)
print(length(mm2Patterns(m2l)))
print(length(mm2ReducedPatterns(m2l)))
}
