% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{find}
\alias{find}
\title{Find the pattern with the best coverage given a set of spectra.}
\usage{
find(m2l, ids, vals = c("P", "S"), ...)
}
\arguments{
\item{m2l}{an ms2Lib Object.}

\item{ids}{A list of ids to be searched for (for now, only spectra are implemented).}

\item{vals}{The type of the returned values (for now, only patterns are implemented (P)).}

\item{...}{supplementary arguments.}
}
\value{
The ids of the patterns or spectra containing all the ids.
}
\description{
Find the pattern with the best coverage, which means that covered the most intensity, for a set of spectra.
}
\examples{
data(m2l)
## The calculation of coverage must have been done before searching
m2l <- calculateCoverage(m2l)
find(m2l, "S7", "P")
}
