% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling.R
\name{findAllComponents}
\alias{findAllComponents}
\title{Find the annotable component of a GNPS network.}
\usage{
findAllComponents(net, minSize = 3, singleClique = 0.9,
  vname = "cluster index")
}
\arguments{
\item{net}{A gnps network}

\item{minSize}{The minimum size of the detected clique.}

\item{singleClique}{Not used at the moment}
}
\value{
A list of the componets to be checked.
}
\description{
Extract the annotated compoents of a network, at the moment, the extracted compoents are limited ot the clique.
}
\examples{
print("Examples to be put here")
}
