% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling_gnps.R
\name{findGNPSComponents}
\alias{findGNPSComponents}
\title{Find the annotable component of a GNPS network.}
\usage{
findGNPSComponents(net, minSize = 3, pairThreshold = 0.9,
  vname = "cluster index", eattr = "cosine_score")
}
\arguments{
\item{net}{A gnps network}

\item{minSize}{The minimum size of the detected clique.}

\item{pariThreshold}{For all the similarities, for each dataset,}
}
\value{
A list of the componets to be checked.
}
\description{
Extract the annotated compoents of a network, at the moment, the extracted compoents are limited ot the clique.
}
\examples{
print("Examples to be put here")
}
