% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\docType{methods}
\name{findMz,ms2Lib-method}
\alias{findMz,ms2Lib-method}
\title{Search in an ms2Lib object?}
\usage{
\S4method{findMz}{ms2Lib}(m2l, mz, type = c("S", "L"), ppm = 15,
  dmz = 0.01)
}
\arguments{
\item{m2l}{ms2Lib object}

\item{mz}{A double giving the mass to be searched.}

\item{type}{The data to be search, "S" for spectra and "L" for losses}

\item{ppm}{The tolerance in ppm}

\item{dmz}{The minimum tolerance in Da, if the ppm tolerance is lower in Da than this threshold, this threshold is selected.}
}
\value{
A character vector giving the IDs of the found losses or elements.
}
\description{
Search an ms2Lib object givena tolerance in ppm or dmz.
}
\examples{
print("examples to be put here")
}
