% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling.R
\name{findPatternsFromComponents}
\alias{findPatternsFromComponents}
\title{Find pattern explainig components}
\usage{
findPatternsFromComponents(m2l, components, metric = c("f1"),
  threshold = NA_real_, simplify = TRUE)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{components}{A list containing the components to be explained.}

\item{metric}{The metric used to match a component and the pattern occurence.}

\item{threshold}{The threshold used to discriminate the dataset}

\item{simplify}{}
}
\description{
Find pattern maximizing the socre givne in metrics with each components. Threshold may be used to indicate that not match
has been found. The majority
}
