% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaGeneration.R
\name{formulaToString}
\alias{formulaToString}
\title{Calculate a string representation of one formula}
\usage{
formulaToString(vformula, vnames = NULL)
}
\arguments{
\item{vformula}{the formula field of an MzDiffFormula object}

\item{vnames}{vector of atom names to consider}
}
\description{
Return the string representation of one formula stored in a MzDiffFormula object.
Should not be used by the user.
}
