% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern.R
\name{mm2Occurrences,fragPattern-method}
\alias{mm2Occurrences,fragPattern-method}
\alias{mm2Root,fragPattern-method}
\alias{mm2Name,fragPattern-method}
\alias{mm2CanonicalForm,fragPattern-method}
\alias{show,fragPattern-method}
\title{Methods for a fragPattern objects}
\usage{
\S4method{mm2Occurrences}{fragPattern}(object)

\S4method{mm2Root}{fragPattern}(object)

\S4method{mm2Name}{fragPattern}(object)

\S4method{mm2CanonicalForm}{fragPattern}(object)

\S4method{show}{fragPattern}(object)
}
\arguments{
\item{object}{A fragPattern object}
}
\value{
None
}
\description{
Methods for a fragPattern objects

Show information about a fragPattern object
}
\examples{
data(m2l)

occs <- mm2Occurrences(m2l["P1"])

root <- mm2Root(m2l["P1"])

name_p <- mm2Name(m2l["P1"])

can_form <- mm2CanonicalForm(m2l["P1"])

show(m2l["P12"])
}
