% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern.R
\name{fragPattern}
\alias{fragPattern}
\title{fragPattern constructor}
\usage{
fragPattern(rlist)
}
\arguments{
\item{rlist}{rlist is a list containg 2 fields,
\begin{itemize}
\item \textbf{edges:} An edge list which includes the "from","to" and "lab" fileds, which by processed
by the graph_from_data_frame function.
\item \textbf{occurences:} A matrix with 2 column, the first giving the id in which the patterns are found
and the second giving the nodes.
\end{itemize}}
}
\value{

}
\description{
This is an internal constructor which is used to pass form the rcpp
output to a fragPattern object including the occurences. It should not be
called directly by the user.
}
\examples{
print("Exmaples to be put here")
}
