% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{get_formula}
\alias{get_formula}
\title{Get all the precursor formulas of the spectra in the dataset}
\usage{
get_formula(m2l)
}
\arguments{
\item{m2l}{a ms2Lib object}
}
\description{
If given as supplementary information, this function returns all
the precursor formulas of the spectra in the dataset
}
\examples{
data(m2l)

get_formula(m2l)
}
