% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzDiffFormula.R
\name{getAtomMass}
\alias{getAtomMass}
\title{Get the mass of an atom}
\usage{
getAtomMass(atoms)
}
\arguments{
\item{atoms}{atom(s) whose mass is searched for. The possible atoms are "C","H", "O", "N", "F", "Cl", "Br", "I", "Si", "S", "P", "Se"}
}
\description{
Get the mass of an atom
}
