% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeLabels.R
\name{getExtendedSpec}
\alias{getExtendedSpec}
\title{Return the spectrum considered by mineMS2 when building the graph.}
\usage{
getExtendedSpec(
  spec,
  ppm = 5,
  dmz = 0.02,
  removeSup = TRUE,
  addPrecursor = TRUE,
  maxNum = NULL,
  relInt = TRUE
)
}
\arguments{
\item{spec}{A "Spectrum2" object as defined in MSnBase}

\item{ppm}{A ppm parameter to match the precursor}

\item{dmz}{A minimum deviation in dalton to match the precursor.}

\item{removeSup}{A boolean indicating if peaks with a mass higher than the precursor should be remode}

\item{addPrecursor}{A boolen indicating if a preucrsor should be added if needed}

\item{maxNum}{The maximum number of peak to keep. The value correspond ot the top-k intensities.}

\item{relInt}{Should the relative intensity be added to the dataset.}
}
\value{
A data.frame containg the mz, intensity and precursor columns.
}
\description{
Return the spectrum considered by mineMS2 when building the graph.
}
\examples{
#Loading data
data(m2l)

getExtendedSpec(m2l["S10"],maxNum = 15)
}
