% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{getInfo}
\alias{getInfo}
\title{Return the available informationon a lost of a spectra.}
\usage{
getInfo(m2l, ids)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{ids}{A vector of IDs should be a spectrum or a losses.}

\item{...}{Supplementary information ot be passed to the getInfo function.}
}
\value{
A data.frame giving informations about the queried elements.
}
\description{
Return the available informationon a lost of a spectra.
}
\examples{
print("Examples to be put here")
}
