% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{getMzDiff}
\alias{getMzDiff}
\title{All the exact m/z differences for a given fragPattern object}
\usage{
getMzDiff(g, occs, mgs, loss_mz)
}
\arguments{
\item{g}{fragPattern object}

\item{occs}{list of occurrences in the fragPattern}

\item{mgs}{list of fragmentation graphs (DAGs) contained in the fragPattern}

\item{loss_mz}{list of all discreized the m/z differences}
}
\description{
Return all the exact m/z differences for a given fragPattern object.
Should not be used by the user.
}
