% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern.R
\name{infoPatterns}
\alias{infoPatterns}
\title{Spectra information for a specific pattern}
\usage{
infoPatterns(m2l, id_p)
}
\arguments{
\item{m2l}{a ms2Lib object}

\item{id_p}{a fragPattern id (such as "P1")}
}
\value{
a data.frame containing information about the spectra containing the pattern
}
\description{
Spectra information for a specific pattern
}
