% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorePatterns.R
\name{kLatticeReduction}
\alias{kLatticeReduction}
\title{Reduce the lattice to k nodes.}
\usage{
kLatticeReduction(m2l, k)
}
\arguments{
\item{m2l}{ms2Lib object.}

\item{k}{The number of remaining nodes.}
}
\value{
the filled ms2Lib object.
}
\description{
Reduce the lattice to the top-k patterns.
}
\examples{
print("Examples here")
}
