% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFormula.R
\docType{methods}
\name{length,LossFormula-method}
\alias{length,LossFormula-method}
\title{Return the number of possible formula for a loss}
\usage{
\S4method{length}{LossFormula}(x)
}
\arguments{
\item{x}{A loss formula object.}
}
\value{
The number of possible formula for a loss.
}
\description{
Return the number of possible formula for a loss. Normally not used by the user.
}
