% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\docType{methods}
\name{length,ms2Lib-method}
\alias{length,ms2Lib-method}
\title{Number of patterns of an mss2Lib object}
\usage{
\S4method{length}{ms2Lib}(x)
}
\arguments{
\item{x}{An m2Lib object to be shown.}
}
\value{
The number of mined patterns.
}
\description{
Return the number of mined patterns of an ms2Lib object
}
\examples{
#Loading the data
data(m2l)
length(m2l)
}
