% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzDiffFormula.R
\name{length,MzDiffFormula-method}
\alias{length,MzDiffFormula-method}
\title{Calculate the number of possible formula for a m/z difference}
\usage{
\S4method{length}{MzDiffFormula}(x)
}
\arguments{
\item{x}{A loss formula object.}
}
\value{
The number of possible formula for a m/z difference.
}
\description{
Return the number of possible formula for a m/z difference. Normally not used by the user.
}
