% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{list_all}
\alias{list_all}
\title{List all patterns ("P") found for each spectrum (S) or loss (l)}
\usage{
list_all(m2l, to_list = c("P"), vals = c("S", "L"), reduced = TRUE)
}
\arguments{
\item{m2l}{an ms2Lib Object.}

\item{to_list}{the type of features to search for (only P for now)}

\item{vals}{the type of features in which we search for the first feature (S or L for now)}

\item{reduced}{if TRUE, we only consider the reduced version of the patterns (with a good coverage)}
}
\value{
A list of the ids of the patterns containing the ith spectrum or loss
}
\description{
List all the patterns ("P") found for each spectrum or containing each loss.
}
\examples{
data(m2l)
#if reduced = TRUE, calculateCoverage must have been executed first
list_all(m2l,"P", "S", reduced=FALSE)
list_all(m2l, vals = "S", reduced=FALSE)
}
