% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{listMzDiffbyPattern}
\alias{listMzDiffbyPattern}
\title{Export m/z differences Table for one pattern}
\usage{
listMzDiffbyPattern(m2l, pattern, golden_rule, spec2Annot, export_pdf)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{pattern}{a fragPattern object}

\item{golden_rule}{a boolean to consider golden rules or not to construct the formula (used for spec2Annot method)}

\item{spec2Annot}{if TRUE, the formula are calculated using spec2Annot (github/odisce)}

\item{export_pdf}{a boolean to export in pdf or not}
}
\value{
A data.frame with the list of m/z differences
}
\description{
Export m/z differences Table with information about one pattern
contains the min, mean and max mz values of losses calculated with all the spectra of the pattern and the loss
formula calculated with the inner algorithm and the package Spec2Annot
}
\examples{
data(m2l)
listMzDiffbyPattern(m2l, m2l["P20"], golden_rule = TRUE, spec2Annot = TRUE, export_pdf = FALSE)
}
