% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{lossesTable}
\alias{lossesTable}
\title{Export Losses Table}
\usage{
lossesTable(m2l)
}
\arguments{
\item{m2l}{An ms2Lib object}
}
\value{
A data.frame with the losses
}
\description{
Export the losses table aith the full informations
}
\examples{
#Loading the data
data(m2l)

head(lossesTable(m2l))
}
