% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2l.R
\docType{data}
\name{m2l}
\alias{m2l}
\title{Example ms2Lib object}
\format{
An object of class \code{ms2Lib} of length 51.
}
\usage{
m2l
}
\description{
The dataset contains 51 MS/MS spectra from secondary metabolites of 
Penicillium nordicum (one spectrum per compound) acquired with a Data 
Dependent Analysis (DDA) approach on a HPLC system (Luna C18 column; 
Phenomenex) coupled to an LTQ Orbitrap XL hybrid (Thermo Fisher Scientific) 
operated in the positive ionization mode at a HCD20 collision energy and a 
resolution of 7,500.
Hautbergue et al. (2019). Combination of Isotope Labeling and Molecular 
Networking of Tandem Mass Spectrometry Data To Reveal 69 Unknown Metabolites 
Produced by Penicillium Nordicum. Analytical Chemistry 91 (19): 12191–202. 
https://doi.org/10.1021/acs.analchem.9b01634.
}
\keyword{datasets}
