% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling.R
\name{mapNetworkMS2lib}
\alias{mapNetworkMS2lib}
\title{Map an MS-MS network to an ms2Lib object.}
\usage{
mapNetworkMS2lib(net, m2l, vname = "cluster index")
}
\arguments{
\item{net}{An igraph object as read by a GNPS netwrk}

\item{m2l}{An ms2Lib object}

\item{vname}{THe name of the vertex attributes used to perform the mapping. It should corrspond to the order of the file in the
"spectraInfo" field of the ms2Lib object, and should be in the same order than the ms2lib library order.}
}
\value{
return a numeric index giving for each spectra in the ms2Lib object the vertex index
}
\description{
Map a network object, an igraph object from a GNPS object to the dataset
}
\examples{
print("Examples to be put here")
}
