% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/ms2Lib.R
\name{mineClosedSubgraphs}
\alias{mineClosedSubgraphs}
\alias{mineClosedSubgraphs,ms2Lib-method}
\title{Extract frequent closed subgraphs.}
\usage{
mineClosedSubgraphs(m2l, ...)

\S4method{mineClosedSubgraphs}{ms2Lib}(m2l, count = 2, sizeMin = 2, precursor = FALSE)
}
\arguments{
\item{m2l}{An ms2Lib object to be processed.}

\item{...}{Supplementary arguments passed to the mining method}

\item{count}{The minimum number of spectra in which the spectrum need to be sampled.}

\item{sizeMin}{The minimum size of the mined patterns.}

\item{precursor}{Should only pattern coming from the root be mined.}
}
\value{
The filled ms2Lib object.
}
\description{
`mineClosedSubgraphs` extracts frequent closed subgraphs from an ms2Lib object.

Mine all the complete closed subgraphs from a set of preconstructed fragmentation graphs objects.
}
\examples{
#Loading the data
data(m2l)

#Mining the subgraphs
m2l <- mineClosedSubgraphs(m2l,count=2,sizeMin = 1)
}
