% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineMS2-package.R
\docType{package}
\name{mineMS2-package}
\alias{mineMS2-package}
\title{Mining of MS-MS spectra by FSM}
\description{
Mining of MS-MS spectra by frequent subgraph mining method.
}
\details{
Mining of MS-MS spectra by frequent subgraph mining method followed. T
}
