% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2l.R
\docType{data}
\name{molnet_df}
\alias{molnet_df}
\title{molecular graph obtained with GNPS for the dataset of MS/MS spectra 
from P. nordicum}
\format{
An object of class \code{list} of length 2.
}
\usage{
molnet_df
}
\description{
molecular graph obtained with GNPS for the dataset of MS/MS spectra 
from P. nordicum
}
\keyword{datasets}
