% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/ms2Lib.R
\name{setIds}
\alias{setIds}
\alias{mm2Spectra,ms2Lib-method}
\alias{mm2SpectraInfos,ms2Lib-method}
\alias{mm2Dags,ms2Lib-method}
\alias{mm2Atoms,ms2Lib-method}
\alias{mm2EdgesLabels,ms2Lib-method}
\alias{mm2Patterns,ms2Lib-method}
\alias{mm2ReducedPatterns,ms2Lib-method}
\alias{setIds,ms2Lib-method}
\title{Methods on ms2lib objects}
\usage{
setIds(m2l, ...)

\S4method{mm2Spectra}{ms2Lib}(m2l)

\S4method{mm2SpectraInfos}{ms2Lib}(m2l)

\S4method{mm2Dags}{ms2Lib}(m2l)

\S4method{mm2Atoms}{ms2Lib}(m2l)

\S4method{mm2EdgesLabels}{ms2Lib}(m2l)

\S4method{mm2Patterns}{ms2Lib}(m2l)

\S4method{mm2ReducedPatterns}{ms2Lib}(m2l)

\S4method{setIds}{ms2Lib}(m2l, ids)
}
\arguments{
\item{m2l}{An m2Lib object.}

\item{...}{Supplementary arguments passed to the setIds method}

\item{ids}{The new ids of an ms2Lib objects}
}
\description{
Different atttributes from a ms2Lib object can be accessed using methods.
}
\details{
For the `setIds` method, changing the ids of spectra can be useful to plot them.
}
\examples{

spectra <- mm2Spectra(m2l)

infos <- mm2SpectraInfos(m2l)

dags <- mm2Dags(m2l)

atoms <- mm2Atoms(m2l)

edge_labels <- mm2EdgesLabels(m2l)

patterns <- mm2Patterns(m2l)

reduced_patterns <- mm2ReducedPatterns(m2l)
#Loading the data
data(m2l)

#Plot with all ids
plotOccurrences(m2l,"P15")

m2l <- setIds(m2l,paste("spectrum n",1:length(mm2Spectra(m2l)),sep=""))

#Plot with the new ids
plotOccurrences(m2l,"P15")
}
