% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzDiffFormula.R
\name{as.character,MzDiffFormula-method}
\alias{as.character,MzDiffFormula-method}
\title{Calculate a string representation of a MzDiffFormula object}
\usage{
\S4method{as.character}{MzDiffFormula}(x)
}
\arguments{
\item{x}{A MzDiffFormula object.}
}
\value{
A character vector with all the possible formula for a loss object.
}
\description{
Return a string representation of all the possible formula. Should not be used by the user.
}
