% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzDiffFormula.R
\name{MzDiffFormulaFromSingleString}
\alias{MzDiffFormulaFromSingleString}
\title{Create a MzDiffFormula object from a string representation of formulas}
\usage{
MzDiffFormulaFromSingleString(formula, ref, sep = "|")
}
\arguments{
\item{formula}{string containing one or several formulas}

\item{ref}{list of atom names included in the formula(s)}

\item{sep}{character separator to distinguish between formulas}
}
\description{
Return a MzDiffFormula object containing all the formulas given in the string in parameter.
Should not be used by the user.
}
\examples{
MzDiffFormulaFromSingleString(formula = "CH6O3|C4H2O,", ref = c("C", "H", "O"), sep = "|")

}
