% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{mzDiffTable}
\alias{mzDiffTable}
\title{Export m/z differences Table}
\usage{
mzDiffTable(m2l)
}
\arguments{
\item{m2l}{An ms2Lib object}
}
\value{
A data.frame with the list of m/z differences
}
\description{
Export the m/z differences table with the full information
}
\examples{
#Loading the data
data(m2l)

head(mzDiffTable(m2l))
}
