% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{mzDiffTableComplete}
\alias{mzDiffTableComplete}
\title{Export m/z differences Table}
\usage{
mzDiffTableComplete(m2l)
}
\arguments{
\item{m2l}{An ms2Lib object}
}
\value{
A data.frame with the list of m/z differences
}
\description{
Export m/z differences Table with information about the patterns containing those m/z differences
for each m/z difference label, contains its id (lab), the mean mz, the minimum mz, the maximum mz, the number of occurrences (count) 
and the possible formula. It also contains the number of patterns in which this label is found (and their ids)
}
\examples{
#Loading the data
data(m2l)

head(mzDiffTableComplete(m2l))
}
