% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeLabels.R
\name{penalizedLossesDefault}
\alias{penalizedLossesDefault}
\title{Default table of penalized loss}
\usage{
penalizedLossesDefault()
}
\value{
A data frame with 2 columns the names of the losses and the
formula of the losses. Only the second column is used the first column is informative.
}
\description{
The losses are penalized by the scoring function,
they are extracted form the paper by Bocker et al 2010. This includes both very
frequent losses, which will lead ot irllevant patterns and forbidden losses.
}
\examples{
penalizedLossesDefault()
}
