% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\name{plot,fragPattern,ANY-method}
\alias{plot,fragPattern,ANY-method}
\title{Plotting a fragPattern object.}
\usage{
\S4method{plot}{fragPattern,ANY}(
  x,
  y = NULL,
  title = NULL,
  edgeLabels = NULL,
  dags = NULL,
  nodeLabel = c("default", "label"),
  edgeLabel = c("formula", "mass", "none"),
  atoms = c("C", "H", "O", "N", "S", "P"),
  formula = NA_character_,
  mzdigits = 3,
  vertex_size = 55,
  vertex_label_cex = 0.8,
  edge_label_cex = 0.8,
  subNodes = NULL,
  tkplot = FALSE,
  print_formula = TRUE,
  colored_edges = c(),
  ...
)
}
\arguments{
\item{x}{The pattern to plot.}

\item{y}{Not used at the moment.}

\item{title}{The title used for the plot}

\item{edgeLabels}{A vector usually passed automatically by the plot method of the ms2Lib object.}

\item{dags}{A list of mass graphs on which the pattern has been calculated.}

\item{nodeLabel}{The type of vertex label to be plotted, default means that will try to be 
determined while label show the raw label as integer.}

\item{edgeLabel}{The type of edges label to be plotted, default means that will try to be 
determined while label show the raw label as integer.}

\item{atoms}{The atoms used in the formula in the right order.}

\item{formula}{The formula of the associated dags.}

\item{mzdigits}{The number of digits included in the plot of m/z differences}

\item{vertex_size}{The size of the vertices.}

\item{vertex_label_cex}{The size of the vertices label}

\item{edge_label_cex}{The size of the edges label}

\item{subNodes}{The subset of nodes ot be plotted if necessary}

\item{tkplot}{Shall an interactive plot be shown using tkplot.}

\item{print_formula}{if TRUE, the possible formula are indicated for edges and vertices}

\item{colored_edges}{list of edges ids to highlight}

\item{...}{Supplementary arguments passed to the igraph plot function}
}
\value{
None
}
\description{
plot a fragPattern using igraph capabilities. This function should not be called directly by the user.
Call `plot(m2l,"x")` instead.
}
\examples{
data(m2l)

plot(m2l, "P12")
}
