% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexPlot.R
\name{plotLattice}
\alias{plotLattice}
\title{Plot the lattice of an ms2Lib object}
\usage{
plotLattice(m2l, reduced = TRUE)
}
\arguments{
\item{m2l}{The ms2Lib object}

\item{reduced}{Shall the reduced lattice or the full lattice be plot.}
}
\description{
Plot the lattice of an ms2Lib object
}
