% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\name{plotOccurences,ms2Lib-method}
\alias{plotOccurences,ms2Lib-method}
\title{Plotting the occurences of a fragPattern object}
\usage{
\S4method{plotOccurences}{ms2Lib}(
  m2l,
  pidx,
  titles = NULL,
  byPage = 6,
  subOccs = NULL,
  highlights = TRUE,
  commonAxis = FALSE,
  ...
)
}
\arguments{
\item{m2l}{An ms2lib object.}

\item{pidx}{A pattern id or a fragPattern object.}

\item{titles}{A vector of titles to be used. A default title includes id
and precursor.}

\item{byPage}{The maximum number of spectra to be plotted by page.}

\item{subOccs}{Shall some specific occurences be considered over all the occurences.}

\item{highlights}{Shall the peaks covered by the pattern be highlighted.}

\item{commonAxis}{Shall all the spectra be plotted with a common x-axis.}

\item{...}{supplementary function to be passed to the plot function.}
}
\value{
A list containng the spectra and their coloring in RGB format
}
\description{
Plot the occurences, the spectra overlayed with the matched peaks of a fragmentation pattern.
}
\examples{
data(m2l)

plotOccurences(m2l, "P12")
}
