% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\name{plotOccurrences,ms2Lib-method}
\alias{plotOccurrences,ms2Lib-method}
\title{Plotting the occurrences of a fragPattern object}
\usage{
\S4method{plotOccurrences}{ms2Lib}(
  m2l,
  pidx,
  titles = NULL,
  byPage = 6,
  subOccs = NULL,
  highlights = TRUE,
  commonAxis = FALSE,
  path_inchi = NULL,
  ggplot.l = TRUE,
  plot.l = TRUE,
  ...
)
}
\arguments{
\item{m2l}{An ms2lib object.}

\item{pidx}{A pattern id or a fragPattern object.}

\item{titles}{A vector of titles to be used. A default title includes id
and precursor.}

\item{byPage}{The maximum number of spectra to be plotted by page.}

\item{subOccs}{Shall some specific occurrences be considered over all the occurrences.}

\item{highlights}{Shall the peaks covered by the pattern be highlighted?}

\item{commonAxis}{Shall all the spectra be plotted with a common x-axis?}

\item{path_inchi}{name of a tabular file containing the inchi keys of the molecules in a column named "name"; if this table is available, the 2D structures corresponding to the spectra will be retrieved from ChemSpider (webchem package) and displayed in the plot along the spectra (default NULL)}

\item{ggplot.l}{Shall ggplot be used?}

\item{plot.l}{Shall the plot be displayed?}

\item{...}{supplementary function to be passed to the plot function.}
}
\value{
A list containing the spectra and their coloring in RGB format
}
\description{
Plot the occurrences, the spectra overlayed with the matched peaks of a fragmentation pattern.
}
\examples{
data(m2l)

plotOccurrences(m2l, "P12")
}
