% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{plotOccurrences}
\alias{plotOccurrences}
\title{Plot occurrences of fragPattern in library object.}
\usage{
plotOccurrences(m2l, pidx, ...)
}
\arguments{
\item{m2l}{A ms2Lib object}

\item{pidx}{A pattern id to be plotted}

\item{...}{Supplementary arguments passed to the plotting method}
}
\description{
`plotOccurrences` Plot a pattern on its associated fragmentation spectrum.
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
