% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexPlot.R
\name{printPDF}
\alias{printPDF}
\title{Export in PDF the information about a pattern}
\usage{
printPDF(m2l, id_p, infos, df_mass_diff, spectra, save_dir)
}
\arguments{
\item{m2l}{ms2Lib object}

\item{id_p}{id of the pattern}

\item{infos}{data frame containing the list of metabolites}

\item{df_mass_diff}{data frame containing the list of m/z differences}

\item{spectra}{list of plots for the spectra}

\item{save_dir}{name of the directory to store the pdf files}
}
\description{
Function to export to PDF the information about a pattern, i.e. list of metabolites containing the pattern, graph of the pattern,
list of m/z differences with possible formula and spectra of the metabolites (with colored peaks for vertices in the pattern)
}
