% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{recognisedFormat}
\alias{recognisedFormat}
\title{Get the supported input forma for mineMS2}
\usage{
recognisedFormat()
}
\value{
A character vector lsiting the supported format
}
\description{
Get the supported input forma for mineMS2
}
\examples{
recognisedFormat()
}
