% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorePatterns.R
\docType{methods}
\name{scoreLosses,ms2Lib-method}
\alias{scoreLosses,ms2Lib-method}
\title{Determine edges scores}
\usage{
\S4method{scoreLosses}{ms2Lib}(m2l, knownLoss = 0.2, selectedLosses = 3,
  monoAtom = -1, weigth = 3.5)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{knownLoss}{THe pnality to use for know losses.}

\item{selectedLosses}{The majorition to use for known loss.}

\item{monoAtom}{THe panality associated to known loss.}

\item{weigth}{The weight associated to the sigmoid function.}
}
\description{
Determine edges scores
}
\examples{
print("exmaples to be put here")
}
