% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorePatterns.R
\docType{methods}
\name{scorePatterns,ms2Lib-method}
\alias{scorePatterns,ms2Lib-method}
\title{Score the patterns contained in an ms2Lib object}
\usage{
\S4method{scorePatterns}{ms2Lib}(m2l)
}
\arguments{
\item{m2l}{An ms2Lib object ot be scored.}
}
\value{
the completed ms2Lib object
}
\description{
Scores the patterns contained in a ms2Lib object using the scoring  of the edges.
shloub b e called after
}
\examples{
print("examples to be put here")
}
