% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\docType{methods}
\name{select,ms2Lib-method}
\alias{select,ms2Lib-method}
\title{Selection function}
\usage{
\S4method{select}{ms2Lib}(m2l, ids, vals = c("P", "S", "L", "F"), ...)
}
\arguments{
\item{m2l}{AN ms2Lib object.}

\item{ids}{Valid IDs of objects.}
}
\value{
The objects of the correspodning type.
}
\description{
Selection function
}
\examples{
print("Examples to be put here")
}
