% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2LibSplit.R
\name{setComponents,ms2LibSplit-method}
\alias{setComponents,ms2LibSplit-method}
\title{Set the components on an ms2LibSplit object.}
\usage{
\S4method{setComponents}{ms2LibSplit}(m2l, components)
}
\arguments{
\item{m2l}{An ms2LibSplit object.}

\item{components}{A list containing the components. Each components needs to be one of the following
\itemize{
\item A vector corresponding to furnished spectra index.
\item A vector of ids corresponding to the ids of the spectra.
\item A vecotr of spectra ids starting with S.
}}
}
\value{
The filled ms2Lib object.
}
\description{
Set the compoentns field of an ms2LibSplit object, whilme checking for their correctness.
}
\examples{
print("examples to be put here")
}
