% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{setComponents}
\alias{setComponents}
\title{Set the componetns of an ms2 library object.}
\usage{
setComponents(m2l, components, ...)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{components}{The components given as a list of integers}

\item{...}{supplementary argument passed to the methods}
}
\description{
`setComponents` Set the components of an ms2Lib object.
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
