% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\docType{methods}
\name{setIds,ms2Lib-method}
\alias{setIds,ms2Lib-method}
\title{Change the id of spectrum of an ms2Lib object}
\usage{
\S4method{setIds}{ms2Lib}(m2l, ids)
}
\arguments{
\item{m2l}{An m2Lib object.}

\item{ids}{The new ids of an ms2Lib objects}
}
\value{
An ms2Lib object with modified ids
}
\description{
Return an ms2Lib object with the new ids
}
\details{
Changing the ids of spectra, which can be useful to plot them.
}
\examples{
#Loading the data
data(m2l)

#Plot with all ids
plotOccurences(m2l,"P15")

m2l <- setIds(m2l,paste("spectrum n",1:length(m2l@spectra),sep=""))

#Plot with the new ids
plotOccurences(m2l,"P15")
}
