% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{setIds}
\alias{setIds}
\title{Set the ids of a MS2lib object}
\usage{
setIds(m2l, ...)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{...}{Supplementary arguments passed to the setIds method}
}
\description{
`setIds` set the ids of an ms2Lib objects
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
