% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern.R
\name{show,fragPattern-method}
\alias{show,fragPattern-method}
\title{Show information about a fragPattern object}
\usage{
\S4method{show}{fragPattern}(object)
}
\arguments{
\item{object}{A fragPattern object}
}
\value{
None
}
\description{
Show information about a fragPattern object
}
\examples{
data(m2l)
show(m2l["P12"])
}
