% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFormula.R
\docType{methods}
\name{show,LossFormula-method}
\alias{show,LossFormula-method}
\title{Show lossFormula object}
\usage{
\S4method{show}{LossFormula}(object)
}
\arguments{
\item{object}{A lossFormula object.}
}
\value{
None
}
\description{
Show the information about a lossFormula object.
}
