% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\docType{methods}
\name{show,ms2Lib-method}
\alias{show,ms2Lib-method}
\title{Show an ms2Lib object.}
\usage{
\S4method{show}{ms2Lib}(object)
}
\arguments{
\item{object}{An m2Lib object to be shown.}
}
\value{
None.
}
\description{
Get the string representation of an ms2Lib object
}
\examples{
#Loading the data
data(m2l)
m2l
}
