% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2LibSplit.R
\name{show,ms2LibSplit-method}
\alias{show,ms2LibSplit-method}
\title{Show an ms2LibSplit object.}
\usage{
\S4method{show}{ms2LibSplit}(object)
}
\arguments{
\item{object}{An m2Lib object to be shown.}
}
\value{
None.
}
\description{
Get the string representation of an ms2LibSplit object
}
