% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\name{spectrum_ggplot}
\alias{spectrum_ggplot}
\title{Plotting the spectrum containing a pattern with the spectrum peaks corresponding to the pattern nodes in color (with ggplot)}
\usage{
spectrum_ggplot(
  i,
  loss_mz,
  mgs,
  g,
  ru_occs_gid,
  u_occs_gid,
  mzprec,
  rtprec,
  names,
  n,
  col_vec,
  path_inchi,
  dir_images,
  x_lim,
  title = NULL
)
}
\arguments{
\item{i}{id of the spectrum in the ms2Lib object}

\item{loss_mz}{all m/z differences in the dataset}

\item{mgs}{all DAGs of the dataset}

\item{g}{graph of the pattern}

\item{ru_occs_gid}{numeric ids of the occurrences of the pattern}

\item{u_occs_gid}{ids of the occurrences of the pattern}

\item{mzprec}{precursor mz of the occurrences of the pattern (as stored in the ms2Lib object)}

\item{rtprec}{precursor rt of the occurrences of the pattern (if given)}

\item{names}{names of the occurrences of the pattern (as stored in the ms2Lib object)}

\item{n}{N ids of the occurrences of the pattern (as stored in the ms2Lib object) (only to create 2D structure images)}

\item{col_vec}{vector of colours for the peaks of the pattern}

\item{path_inchi}{name of a tabular file containing the inchi keys of the molecules in a column named "name"; if this table is available, the 2D structures corresponding to the spectra will be retrieved from ChemSpider (webchem package) and displayed in the plot along the spectra (default NULL)}

\item{dir_images}{path to the directory to store the png images}

\item{x_lim}{the limits in mz values for the spectra of the pattern}

\item{title}{the title for the spectra. If NULL, the title will be set automatically}
}
\value{
ggplot object
}
\description{
Plotting the spectrum containing a pattern with the spectrum peaks corresponding to the pattern nodes in color (with ggplot)
}
