% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFormula.R
\name{[,LossFormula,ANY,ANY,ANY-method}
\alias{[,LossFormula,ANY,ANY,ANY-method}
\title{Selecting a fornula among the different loss formula possible.}
\usage{
\S4method{[}{LossFormula,ANY,ANY,ANY}(x, i, j = NULL, ..., drop = TRUE)
}
\arguments{
\item{x}{The Loss formula object}

\item{i}{The position of the selected formula}

\item{j}{Not used}

\item{...}{Not used}

\item{drop}{Not used}
}
\value{
The formula as 1xatoms matrix
}
\description{
Selecting a fornula among the different loss formula possible.
}
\examples{
print("Example to be put here")
}
