% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{[,ms2Lib,ANY,ANY,ANY-method}
\alias{[,ms2Lib,ANY,ANY,ANY-method}
\title{Indexing function}
\usage{
\S4method{[}{ms2Lib,ANY,ANY,ANY}(x, i, j = NULL, ..., drop = TRUE)
}
\arguments{
\item{x}{An ms2Lib oject.}

\item{i}{The index, a string starting by S if it is a spectrum, P if it is a pattern D if it is a DAG, L if it is a m/z difference label}

\item{j}{unused.}

\item{...}{unused.}

\item{drop}{unused.}
}
\value{
A list containing the object.
}
\description{
Return an object given its index: can be a spectrum, a pattern, a DAG or a m/z difference.
}
\examples{
#Loading the data
data(m2l)

#Extracting a spectrum
m2l["S10"]

#The associated DAG
m2l["D10"]

#Extraction of a pattern
m2l["P54"]

#Extraction of m/z difference information
m2l["L20"]
}
