% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzDiffFormula.R
\name{[,MzDiffFormula,ANY,ANY,ANY-method}
\alias{[,MzDiffFormula,ANY,ANY,ANY-method}
\title{Selecting a formula among the different loss formula possible.}
\usage{
\S4method{[}{MzDiffFormula,ANY,ANY,ANY}(x, i, j = NULL, ..., drop = TRUE)
}
\arguments{
\item{x}{The Loss formula object}

\item{i}{The position of the selected formula}

\item{j}{Not used}

\item{...}{Not used}

\item{drop}{Not used}
}
\value{
The formula as 1xatoms matrix
}
\description{
Selecting a formula among the different loss formula possible.
}
