% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\name{support_ggplot}
\alias{support_ggplot}
\title{Plotting the spectra occurrences of a pattern, i.e. the spectra containing the fragPattern object, with the peaks matching the pattern in color (ggplot version)}
\usage{
support_ggplot(m2l, pidx, titles = NULL, path_inchi = NULL)
}
\arguments{
\item{m2l}{ms2lib object.}

\item{pidx}{pattern id or a frag_pattern object.}

\item{titles}{list of titles for the spectra. If NULL, a title will be set automatically}

\item{path_inchi}{name of a tabular file containing the inchi keys of the molecules in a column named "name"; if this table is available, the 2D structures corresponding to the spectra will be retrieved from ChemSpider (webchem package) in the same directory as path_inchi, and displayed in the plot along the spectra (default NULL)}
}
\value{
A list of plots, one for every spectrum
}
\description{
Plot the spectra occurrences of a pattern, with the matched peaks in color.
}
\examples{
data(m2l)
support_ggplot(m2l, "P12")
}
