% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\docType{methods}
\name{vrange,ms2Lib-method}
\alias{vrange,ms2Lib-method}
\title{Return the range of iteration on an ms2Lib object.}
\usage{
\S4method{vrange}{ms2Lib}(m2l, type = c("S", "L", "P", "F"),
  reduced = TRUE, as.number = FALSE)
}
\arguments{
\item{m2l}{AN ms2Lib object}

\item{type}{"S","L" or "P"}

\item{reduced}{Used only if "type" is set to "P", shall the filtered pattern set be returned.}

\item{as.number}{If as number is selected integer are returned without the prefix.}
}
\value{
A character vector giving the existing ids in the ms2Lib object,
}
\description{
Return the range of iteration on an ms2Lib object.
}
\examples{
print("Examples to be put here")
}
