% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{vrange}
\alias{vrange}
\title{Set the components of an ms2 library object.}
\usage{
vrange(m2l, type, ...)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{type}{The type of range to be returned}

\item{...}{supplementary argument passed to the methods}
}
\description{
`vrange` return the range of the different elements of an ms2Lib object.
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
