% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling_gnps.R
\name{annotateNetwork}
\alias{annotateNetwork}
\title{Annotation of a GNPS network with patterns for each detected component}
\usage{
annotateNetwork(
  components,
  net,
  patterns,
  copy = TRUE,
  keepattr = c("mz", "rt", "name", "shared name", "formula"),
  sep_infos = ","
)
}
\arguments{
\item{components}{The components of the GNPS network as integers.}

\item{net}{The GNPS network as an igraph object}

\item{patterns}{The set of patterns extracted from an ms2Lib object.}

\item{copy}{Shall the igraph object be copied (recommended)}

\item{keepattr}{Which attributes shall be kept on the network}

\item{sep_infos}{The separator used to separate the colors when multiple colors are associated ot a single node.}
}
\value{
The annotated network as an igraph object.
}
\description{
Annotation of a GNPS network with patterns for each detected component
}
\examples{
library(igraph)
data(m2l)
data(molnet_df)
molnet_igraph <- graph_from_data_frame(molnet_df, directed = FALSE, vertices = NULL)
components <- findGNPSComponents(molnet_igraph)
patterns <- findPatternsExplainingComponents(m2l, components)
molnet_annotated <- annotateNetwork(components, molnet_igraph, patterns)
}
