% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling_gnps.R
\name{annotateNetwork}
\alias{annotateNetwork}
\title{Annotation of GNPS network}
\usage{
annotateNetwork(
  components,
  net,
  patterns,
  copy = TRUE,
  keepattr = c("Compound_Name", "SpectrumID", "DefaultGroups", "cluster index",
    "precursor mass"),
  sep_infos = ","
)
}
\arguments{
\item{components}{The components of the GNPS network as integer.}

\item{net}{The GNPS network as an igraph object}

\item{patterns}{The set of patterns extract from an ms2Lib object.}

\item{copy}{Shall the igraph object be copied (recommended)}

\item{keepattr}{Which attributes sahll be kept on the ntwork}

\item{sep_infos}{The seprator used to separate the colors when multiple colors are associated ot a single node.}
}
\value{
The annotated netowrk.
}
\description{
Annotation of GNPS network
}
\examples{
print("Examples to be put here")
}
