% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{calculateCoverage,ms2Lib-method}
\alias{calculateCoverage,ms2Lib-method}
\title{Calculate the coverage}
\usage{
\S4method{calculateCoverage}{ms2Lib}(x)
}
\arguments{
\item{x}{The ms2Lib to be computed.}
}
\value{
The m2l object with all the coverage calculated.
}
\description{
Calculate the coverage, which means the total intensity covered by the patterns on the mass graphs.
This calculation can be quite long.
}
\examples{
#Loading the data
data(m2l)

#Calculate the coverage for an ms2Lib object
m2l <- calculateCoverage(m2l)
}
