% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{calculateCoverage}
\alias{calculateCoverage}
\title{Calculate the coverage of all the patterns in the dataset.}
\usage{
calculateCoverage(m2l)
}
\arguments{
\item{m2l}{The ms2Lib to bo computed.}
}
\value{
The m2l object with all the coverage calculated.
}
\description{
Calculate the coverage of all the patterns in the dataset.
}
\examples{
print("Examples to be put here")
}
