% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{calculateCoverage}
\alias{calculateCoverage}
\title{Calculate the relative intensity covered by patterns.}
\usage{
calculateCoverage(x, ...)
}
\arguments{
\item{x}{An ms2Lib of a fragPattern object}

\item{...}{supplementary argument passed to the methods}
}
\description{
`calculateCoverage` calculate the relative intensity covered by a pattern or a library of patterns.
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
