% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/discretizeLabels.R
\name{discretizeMzDifferences}
\alias{discretizeMzDifferences}
\alias{discretizeMzDifferences,ms2Lib-method}
\title{Discretize m/z differences.}
\usage{
discretizeMzDifferences(m2l, ...)

\S4method{discretizeMzDifferences}{ms2Lib}(
  m2l,
  ppm = 7,
  dmz = 0.002,
  count = 2,
  limMzFormula = c(14.5, 200),
  maxFrags = 15,
  maxOverlap = 0.05,
  strictMatching = TRUE,
  precPpm = 20,
  precDmz = 0.02,
  atoms = NULL,
  heteroAtoms = TRUE,
  mzDigits = 4
)
}
\arguments{
\item{m2l}{The ms2 lib object to be discretized.}

\item{...}{Supplementary arguments passed to the discretization method}

\item{ppm}{the maximum authorized deviation in ppm (parts per million).}

\item{dmz}{The maximum authorized deviation in Da.}

\item{count}{The minimum number of spectra in which the label needs to be found. Shall be greater or equal to 2.}

\item{limMzFormula}{An interval giving the range in which the formula will be calculated.
peaks with a m/z lower than the lower term will be ignored while peak with a mass higher than the
higher term won't have formula check.}

\item{maxFrags}{The maximum number of fragment allowed on the spectra.}

\item{maxOverlap}{The degree of overlap allowed between bins. Bins which overlap more are fused.}

\item{strictMatching}{Shall the formula matching be strict, or approximated.}

\item{precPpm}{The ppm tolerance used ot match the precursor mz to the fragments.}

\item{precDmz}{The minimum tolerance used to match the precursor mz to the fragments.}

\item{atoms}{A list of the maximum number of atoms authorized during the formula generation process.
If it is NULL, the default list the maximum of \emph{limMzFormula}/12, H:50, N:6 ,O:6.}

\item{heteroAtoms}{a boolean,  ignored if a custom atoms is furnished, add heteroatoms P, Cl, S as possible atoms.}

\item{mzDigits}{The number of decimals to write as an information in the graphs.}
}
\value{
An ms2Lib object with filled fields and constructed dags.
}
\description{
`discretizeMzDifferences` discretizes the m/z differences in MS/MS spectra.

It also builds fragmentation graphs from MS/MS spectra using discretized m/z differences as edges.
}
\examples{
#Loading the data
data(m2l)

#Constructing edge labels
m2l <- discretizeMzDifferences(m2l,ppm=15,dmz=0.007,count=2,
precPpm=20,precDmz=0.03,maxFrags=15)

#Constructing edge labels while including heteroatoms
m2l <- discretizeMzDifferences(m2l,atoms = 
list("C"=50,"H"=100,"N"=6,"O"=6,"S"=2,"Cl"=1,"P"=2))
}
