% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filterPatterns}
\alias{filterPatterns}
\title{Title}
\usage{
filterPatterns(
  m2l,
  filter_function = function(x, threshold, num) {     return(sum(x@occurences[,
    "coverage"] > threshold) >= num) },
  args = list(threshold = 0.2, num = 2)
)
}
\arguments{
\item{m2l}{The ms2Lib object ot be filtered}

\item{filter_function}{The filtering function this function take as input a pattern and return TRUE or FALSE.}

\item{args}{}
}
\value{
The supplementary ms2Lib arguments.
}
\description{
Title
}
\examples{
print("Exmaples to be put here.")
}
