% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{find.patterns.class}
\alias{find.patterns.class}
\title{Find biggest F1 score}
\usage{
find.patterns.class(
  m2l,
  ids,
  type = c("f1", "precision", "size"),
  returnall = FALSE,
  full = TRUE,
  reduced = FALSE
)
}
\arguments{
\item{m2l}{An ms2Lib object.}

\item{ids}{Valid IDs of objects, it mays be a chracter vector with the prefix "P" or an integer or numeric vector}

\item{type}{Which criteria is used to determine the best match, F-score, precision or acccuracy}

\item{returnall}{Shall all the patterns with a similar F1 score be returned.}

\item{full}{Shall only full matches be returned.}

\item{reduced}{Shall only the reduced set of patterns be considered.}
}
\value{
A data.frame containing two slots, id the ids of the elements and f1_score the f1 score of the elements
}
\description{
Given a list of molecules ids as a character or integer input, find the pattern maximizing the F1 score of
the molecules.
}
\details{
The computation of this F1 score.
}
\examples{
print("Examples to be put here")
}
