% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{find}
\alias{find}
\title{find patterns or spectra form a set of pattern or spectra ids}
\usage{
find(m2l, ids, vals = c("P", "S"), ...)
}
\arguments{
\item{m2l}{an ms2Lib Object.}

\item{ids}{A list of ids to be searched for.}

\item{vals}{The type of the returned values.}

\item{...}{supp arguments.}
}
\value{
The ids of the patterns or spectra cotaining all the ids.
}
\description{
find data of type "vals" containing all the ids listed in "ids"
}
\examples{
print("examples to be put here")
}
