% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling_gnps.R
\name{findGNPSComponents}
\alias{findGNPSComponents}
\title{Extract the components of a molecular network}
\usage{
findGNPSComponents(
  net,
  minSize = 3,
  pairThreshold = 0.9,
  vname = "cluster index",
  eattr = "cosine_score"
)
}
\arguments{
\item{net}{A GNPS network}

\item{minSize}{The minimum size of the detected clique.}

\item{pairThreshold}{A threshold used to discard the edges to detect components.}

\item{vname}{The name of the exported index of the component as vertex attribute.}

\item{eattr}{The name of the considered similarity measure on the network.}
}
\value{
A list of the components to be checked.
}
\description{
Extract the components of a molecular network (the cliques, the connected components and the high similarity pairs)
}
\examples{
data(molnet_df)
molnet_igraph <- igraph::graph_from_data_frame(molnet_df$edges, directed = FALSE, vertices = molnet_df$vertices)
list_components <- findGNPSComponents(molnet_igraph, minSize = 3, pairThreshold = 0.9)
print(list_components)
}
