% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling_gnps.R
\name{findGNPSComponents}
\alias{findGNPSComponents}
\title{Find the annotable component of a GNPS network.}
\usage{
findGNPSComponents(net, minSize = 3, pairThreshold = 0.9,
  vname = "cluster index", eattr = "cosine_score",
  convert_component = TRUE)
}
\arguments{
\item{net}{A gnps network}

\item{minSize}{The minimum size of the detected clique.}

\item{pairThreshold}{A threshold used to discard the edges to detect components.}

\item{vname}{The name of the exported index of the component as verte attribute.}

\item{eattr}{The name of the considered similarity measure on the network.}
}
\value{
A list of the componets to be checked.
}
\description{
Extract the annotated compoents of a network, at the moment, the extracted compoents are limited ot the clique.
}
\examples{
print("Examples to be put here")
}
