% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling.R
\name{findPatternsExplainingComponents}
\alias{findPatternsExplainingComponents}
\title{Find pattern explainig components}
\usage{
findPatternsExplainingComponents(m2l, components, metric = c("f1"),
  ref_label = NULL, threshold = NA_real_, reduced = TRUE)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{components}{A list containing the components to be explained.}

\item{metric}{The metric used to match a component and the pattern occurence.}

\item{ref_label}{A set of reference label to be used if the extracted component corresponds to an id over the position ofht e MS-MS spetra in the file}

\item{threshold}{The threshold used to discriminate the dataset}
}
\value{
A list containing one data.frame by component including the following informations:
  \code{id} The id of the matched patterns.
  \code{f1} The F1-score between the pattern and the component.
  \code{accuracy} The accuracy between the pattern and the component.
  \code{recall} The recall between the pattern and the component.
  \code{miss} The miss rate between the pattern and the component.
}
\description{
Find pattern maximizing the socre givne in metrics with each components. Threshold may be used to indicate that not match
has been found. The majority
}
\examples{
print("Examples to be put here")
}
