% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpassCoupling.R
\name{findPatternsExplainingComponents}
\alias{findPatternsExplainingComponents}
\title{Find patterns maximizing a score for each component}
\usage{
findPatternsExplainingComponents(
  m2l,
  components,
  metric = c("f1"),
  ref_label = NULL,
  threshold = NA_real_,
  reduced = TRUE,
  top = 1
)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{components}{A list containing the components to be explained.}

\item{metric}{The metric or list of metrics used to match a component and the pattern occurrence.}

\item{ref_label}{A set of reference label to be used if the extracted component corresponds to an id over the position of the MS-MS spetra in the file}

\item{threshold}{The threshold used to discriminate the dataset}

\item{reduced}{Shall only the filtered patterns be considered}

\item{top}{the number of best patterns to return (by default 1)}
}
\value{
A list containing one data.frame by component including the following informations:
  \code{id} The id of the matched patterns.
  \code{f1} The F1-score between the pattern and the component.
  \code{precision} The precision between the pattern and the component.
  \code{recall} The recall between the pattern and the component.
  \code{miss} The miss rate between the pattern and the component.
	 \code{size} The number of vertices in the graph pattern
}
\description{
Find patterns maximizing the score given in metrics for each component.
Threshold may be used to indicate that no match has been found.
Several best patterns can be returned (argument top)
}
\examples{
#Loading the data
data(m2l)
data(molnet_df)

#Finding explainig component 
molnet_igraph <- igraph::graph_from_data_frame(molnet_df$edges, directed = FALSE, vertices = molnet_df$vertices)
components <- findGNPSComponents(molnet_igraph)
print(length(components))
fp <- findPatternsExplainingComponents(m2l,components)
print(fp[[1]])
#Plotting the first explaining pattern
plot(m2l,fp[[1]][1,"id"])
}
