% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/fragPattern.R
\name{mm2Graph}
\alias{mm2Graph}
\alias{mm2Occurrences}
\alias{mm2Root}
\alias{mm2Name}
\alias{mm2CanonicalForm}
\alias{mm2Graph,fragPattern-method}
\alias{mm2Occurrences,fragPattern-method}
\alias{mm2Root,fragPattern-method}
\alias{mm2Name,fragPattern-method}
\alias{mm2CanonicalForm,fragPattern-method}
\alias{show,fragPattern-method}
\title{Show information about a fragPattern object}
\usage{
mm2Graph(object)

mm2Occurrences(object)

mm2Root(object)

mm2Name(object)

mm2CanonicalForm(object)

\S4method{mm2Graph}{fragPattern}(object)

\S4method{mm2Occurrences}{fragPattern}(object)

\S4method{mm2Root}{fragPattern}(object)

\S4method{mm2Name}{fragPattern}(object)

\S4method{mm2CanonicalForm}{fragPattern}(object)

\S4method{show}{fragPattern}(object)
}
\arguments{
\item{object}{A fragPattern object}
}
\value{
None
}
\description{
Show information about a fragPattern object
}
\examples{
data(m2l)
## igraph object to store the vertices and edges of a pattern
graph <- mm2Graph(m2l["P1"])

occs <- mm2Occurrences(m2l["P1"])

root <- mm2Root(m2l["P1"])

name_p <- mm2Name(m2l["P1"])

can_form <- mm2CanonicalForm(m2l["P1"])

show(m2l["P12"])
}
