% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern.R
\name{mm2Graph}
\alias{mm2Graph}
\alias{fragPattern-mm2Graph}
\alias{fragPattern}
\title{fragPattern constructor}
\usage{
\S4method{mm2Graph}{fragPattern}(pat)

fragPattern(rlist)
}
\arguments{
\item{rlist}{rlist is a list containg 2 fields,
\code{edges}: An edge list which includes the "from","to" and "lab" fields, which be processed
by the graph_from_data_frame function.
\code{occurrences}: A matrix with 2 columns, the first giving the id in which the patterns are found
and the second giving the nodes.}
}
\value{
The constructed fragPattern object.
}
\description{
This is an internal constructor which is used to pass from the rcpp
output to a fragPattern object including the occurrences. It should not be
called directly by the user.
}
\examples{
data(m2l)
## igraph object to store the vertices and edges of a pattern
graph <- mm2Graph(m2l["P1"])
}
