% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{getInfo}
\alias{getInfo}
\title{Return the available informations on a component of in an ms2Lib object.}
\usage{
getInfo(m2l, ids)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{ids}{A vector of IDs should be a spectrum or a losses.}

\item{...}{Supplementary information ot be passed to the getInfo function.}
}
\value{
A data.frame giving informations about the queried elements.
}
\description{
Return the available informations on a component of in an ms2Lib object.
}
\examples{
#Loading the data
data(m2l)

#Infos on 2 spectra
getInfo(m2l,c("S10","S25"))

#Infos on multiple losses
getInfo(m2l,c("L12","L42"))

#Example combined with the findMz 
getInfo(m2l,findMz(m2l,391.1398,dmz=0.01))
}
