% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{getInfo}
\alias{getInfo}
\title{Available information}
\usage{
getInfo(m2l, ids)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{ids}{A vector of IDs should be spectra or a m/z differences.}

\item{...}{Supplementary information to be passed to the getInfo function.}
}
\value{
A data.frame giving information about the queried elements.
}
\description{
Return the available information on an element of an ms2Lib object.
}
\examples{
#Loading the data
data(m2l)

#Infos on 2 spectra
getInfo(m2l,c("S10","S25"))

#Infos on all the spectra
getInfo(m2l, "S")

#Infos on multiple losses
getInfo(m2l,c("L12","L42"))

#Example combined with the findMz 
getInfo(m2l,findMz(m2l,391.1398,dmz=0.01))
}
