% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{mineClosedSubgraphs,ms2Lib-method}
\alias{mineClosedSubgraphs,ms2Lib-method}
\title{Mine recurrent subgraph from a set of graphs.}
\usage{
\S4method{mineClosedSubgraphs}{ms2Lib}(m2l, count = 2, sizeMin = 2, precursor = FALSE)
}
\arguments{
\item{m2l}{An m2Lib object to be processed.}

\item{count}{The number of spectra in which the spectrum need to be sampled.}

\item{sizeMin}{The minimum size of the mined patterns.}

\item{precursor}{Should only the occurences coming form the root be conserved.}
}
\value{
The filled ms2Lib object.
}
\description{
Mine all the complete clsoed recurring subgraphs from a set of preconstructed mass graphs objects.
}
\examples{
print("examples to be put here")
}
