% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{mineClosedSubgraphs}
\alias{mineClosedSubgraphs}
\title{Extract frequent closed subgraphs.}
\usage{
mineClosedSubgraphs(m2l, ...)
}
\arguments{
\item{m2l}{A ms2Lib object}

\item{...}{Supplmentary arguments passed to the mining method}
}
\description{
`mineClosedSubgraphs` Extract frequent closed subgraphs from an ms2Lib object.
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
