% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/ms2Lib.R
\name{mineClosedSubgraphs}
\alias{mineClosedSubgraphs}
\alias{mineClosedSubgraphs,ms2Lib-method}
\title{Extract frequent closed subgraphs.}
\usage{
mineClosedSubgraphs(m2l, ...)

\S4method{mineClosedSubgraphs}{ms2Lib}(m2l, count = 2, sizeMin = 1, precursor = FALSE)
}
\arguments{
\item{m2l}{ms2Lib object to be processed}

\item{...}{Supplementary arguments passed to the mining method}

\item{count}{Minimum number of spectra in which a pattern is found; must be at least 2 [default: 2]}

\item{sizeMin}{Minimum size of the mined patterns [default: 1]}

\item{precursor}{Should only patterns coming from the root be mined? [default: FALSE]}
}
\value{
The updated ms2Lib object with the mined patterns.
}
\description{
`mineClosedSubgraphs` extracts frequent closed subgraphs from an ms2Lib object.

Mine all the complete closed subgraphs from a set of preconstructed fragmentation graphs objects.
}
\examples{
#Loading the data
data(m2l)

#Mining the subgraphs
m2l <- mineClosedSubgraphs(m2l)
}
