% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ms2Lib-class}
\alias{ms2Lib-class}
\alias{ms2Lib}
\title{A class storing a set of MS-MS spectra, their associated graphs,
and some supplemnetary inofmration.}
\description{
A class storing a set of MS-MS spectra, their associated graphs,
and some supplemnetary inofmration.
}
\section{Slots}{

\describe{
\item{\code{spectraInfo}}{A data.frame including the spectra information at minima the "mz.precursor" field
and informations added by the user eventually.}

\item{\code{spectra}}{A list of spectra sotred under the form of Spectrum2 object.}

\item{\code{dags}}{A list storing the set of graphs object corresponding to the MS-MS spectra.}

\item{\code{edges.labels}}{The discretized edges labels.}

\item{\code{nodes.labels}}{The discretized nodes labels.}
}}

