% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{ms2Lib}
\alias{ms2Lib}
\title{ms2Lib constructor}
\usage{
ms2Lib(x, suppInfos = NULL)
}
\arguments{
\item{x}{May be one of the following
\itemize{
\item A character vector giving the path to a drictory full of readable format.
\item A list of spectrum2 object which will be integrated directly.
\item A single .mgfspectrum regrouping multiple files.
}}

\item{suppInfos}{Supplementary information to be associated to the spectra.
It should be of the same size as the number of spectra. If there is a "file" column, this
column is used to match the file.}
}
\description{
Create a ms2Lib object from files in suppoerted input formats.
}
\examples{
print("examples to be put here")
}
