% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{ms2Lib}
\alias{ms2Lib}
\title{ms2Lib constructor}
\usage{
ms2Lib(x, suppInfos = NULL, ids = NULL, intThreshold = NULL,
  infosFromFiles = FALSE)
}
\arguments{
\item{x}{May be one of the following
\itemize{
\item A character vector giving the path to a drictory full of readable format.
\item A list of spectrum2 object which will be integrated directly.
\item A single .mgfspectrum regrouping multiple files.
}}

\item{suppInfos}{Supplementary information to be associated to the spectra.
It should be of the same size as the number of spectra. If there is a "file" column, this
column is used to match the filenames. A "composition" or "formula" fields may also be present, and is then used to set the
formula of the molecules.}

\item{ids}{A supplementary vector giving a set of ids to design the spectra. It may be any character vector which
does not start with \code{'P', 'L', 'S'} as they are used internally by mineMS2. Alternatively if a suppInfos table is furnished
and it contains an id fields, it will be used. If no ids are furnished, an id will be generated for each spectra in the form \code{'S1', 'S2', ..., 'SN'}
where N is the number of furnished spectra.}

\item{infosFromFiles}{Shall the other informations present in the files be added to the supplementary infos.}
}
\description{
Create a ms2Lib object from file sin the correct format. Eventually add some supplementary informations given in a file. These supplementary informations
may notably include the composition or the ofrmula of the molecules, which will then be used to plot the created graphs..
}
\examples{
print("examples to be put here")
}
