% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\docType{methods}
\name{plot,fragPattern,ANY-method}
\alias{plot,fragPattern,ANY-method}
\title{plotting a fargPattern object.}
\usage{
\S4method{plot}{fragPattern,ANY}(x, y = NULL, title = NULL,
  edgeLabels = NULL, nodeLab = c("default", "label"),
  edgeLab = c("adapted", "formula", "none"), atoms = c("C", "H", "O",
  "N", "S", "P"), formula = NA_character_, mzdigits = 3,
  vertex_size = 55, subNodes = NULL, tkplot = FALSE, ...)
}
\arguments{
\item{y}{= NULL,}

\item{edgeLabels}{= NULL,}

\item{nodeLab}{= c("default", "label"),}

\item{edgeLab}{= c("formula","none"),}

\item{mzdigits}{= 3,}

\item{vertex_size}{= 30,}

\item{tkplot}{= FALSE,}

\item{...}{}
}
\value{
nothing
}
\description{
plot a fragPattern associated fragmentation
dag using igraph capabilities.
}
\examples{
print("Exmaples ot be put here.")
}
