% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\docType{methods}
\name{plot,fragPattern,ANY-method}
\alias{plot,fragPattern,ANY-method}
\title{plotting a fargPattern object.}
\usage{
\S4method{plot}{fragPattern,ANY}(x, y = NULL, title = NULL,
  edgeLabels = NULL, dags = NULL, nodeLabel = c("default", "label"),
  edgeLabel = c("formula", "mass", "none"), atoms = c("C", "H", "O",
  "N", "S", "P"), formula = NA_character_, mzdigits = 3,
  vertex_size = 55, vertex_label_cex = 0.7, edge_label_cex = 0.7,
  subNodes = NULL, tkplot = FALSE, ...)
}
\arguments{
\item{x}{The pattern to plot.}

\item{y}{Not used at the moment.}

\item{title}{The title used for th eplot}

\item{edgeLabels}{A vector usually passed automatically by the plot method of the ms2Lib object.}

\item{dags}{A list of mass graphs on which the patterns has been calculated.}

\item{nodeLabel}{The type of vertex label to be plotted, default means that a will try to be 
determined while label show the raw label as integer.}

\item{edgeLabel}{The type of edges label to be plotted, default means that a will try to be 
determined while label show the raw label as integer.}

\item{atoms}{The used in the formula in the right order.}

\item{formula}{The formula of the associated dags.}

\item{mzdigits}{The number of digits included in the plot of mass differences}

\item{vertex_size}{The size of the vertices.}

\item{vertex_label_cex}{The size of the vertices label}

\item{edge_label_cex}{The size of the edges label}

\item{subNodes}{The subset of nodes ot be plotted if necessary}

\item{tkplot}{Shall an interactiv eplot be shown using tkplot.}

\item{...}{Supplmentary arguments passed to the igraph plot function}
}
\value{
None
}
\description{
plot a fragPattern associated fragmentation
dag using igraph capabilities. This function should not be called directly by the user.
Call `plot(m2l,"x")` instead.
}
\examples{
print("Examples ot be put here.")
}
