% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\docType{methods}
\name{plot,fragPattern,ANY-method}
\alias{plot,fragPattern,ANY-method}
\title{plotting a fargPattern object.}
\usage{
\S4method{plot}{fragPattern,ANY}(x, y = NULL, title = NULL,
  edgeLabels = NULL, dags = NULL, nodeLabel = c("default", "label"),
  edgeLabel = c("formula", "mass", "none"), atoms = c("C", "H", "O",
  "N", "S", "P"), formula = NA_character_, mzdigits = 3,
  vertex_size = 55, vertex_label_cex = 0.7, subNodes = NULL,
  tkplot = FALSE, ...)
}
\arguments{
\item{x}{The pattern to plot.}

\item{y}{= NULL}

\item{edgeLabels}{= NULL,}

\item{mzdigits}{= 3,}

\item{vertex_size}{= 30,}

\item{vertex_label_cex}{= 0.7,}

\item{subNodes}{= NULL,}

\item{tkplot}{= FALSE,}

\item{...}{Supplmentary arguments passed to the igraph function}

\item{nodeLab}{= c("default", "label"),}

\item{edgeLab}{= c("formula","none"),}
}
\value{
None
}
\description{
plot a fragPattern associated fragmentation
dag using igraph capabilities.
}
\examples{
print("Exmaples ot be put here.")
}
