% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\docType{methods}
\name{plot,ms2Lib,ANY-method}
\alias{plot,ms2Lib,ANY-method}
\title{Plot an element given an idx}
\usage{
\S4method{plot}{ms2Lib,ANY}(x, y, title = NULL, ...)
}
\arguments{
\item{x}{An ms2Lib oject.}

\item{y}{The index, a string starting by S if it a spectrum, P if it's a pattern or D if it's a dag.}

\item{title}{The title of the plot, only used for dag and spectrum.}

\item{...}{supplementary arguments to be passed by the method.}
}
\value{
a fragPattern object of an igraph graph object.
}
\description{
The method depends of the type of the ID furnished. The following prefixes are supported:
\code{'P'} A pattern is called plot method of fragPattern object.
\code{'S'} A spectrum is plotted calling the Plot method of spectrum 2 object.
Any other value will be removed.
}
\examples{
#' #Loading the data
data(m2l)

#Plotting a pattern
plot(m2l,"S10")

#The associated DAG
plot(m2l,"D10")

#Plotting a pattern
plot(m2l,"P53")
}
