% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\name{plotOccurences,ms2Lib-method}
\alias{plotOccurences,ms2Lib-method}
\title{PLotting the occurences of a fragPattern object}
\usage{
\S4method{plotOccurences}{ms2Lib}(
  m2l,
  pidx,
  titles = NULL,
  byPage = 6,
  subOccs = NULL,
  highlights = TRUE,
  commonAxis = FALSE,
  ...
)
}
\arguments{
\item{m2l}{An ms2lib object.}

\item{pidx}{A pattern id or a frag_pattern object.}

\item{titles}{A vector of titles to be used. A default tile include id
and precursor will bne used by default}

\item{byPage}{The maximum number of spectra to be plotted by page.}

\item{subOccs}{Shall some specific occurences be ocnsidered over all the occurences.}

\item{highlights}{Shall the peaks covered by the pattern be highlighted.}

\item{commonAxis}{Shall all the spectra be plotted with a common x-axis.}

\item{...}{supplementary function to be passed to theplot function.}
}
\value{
A list ocntainng the spectra and their coloring in RGB format
}
\description{
Plot the occurences, the spectra overlayed with the matched peaks of a fragmentation pattern.
}
\examples{
print("Examples to be put here")
}
