% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPattern-plot.R
\docType{methods}
\name{plotOccurences,ms2Lib-method}
\alias{plotOccurences,ms2Lib-method}
\title{PLotting the occurences of a fragPattern object inside
an ms2Lib object.}
\usage{
\S4method{plotOccurences}{ms2Lib}(m2l, pidx, titles = NULL, byPage = 6,
  subOccs = NULL, highlights = TRUE, commonAxis = FALSE, ...)
}
\arguments{
\item{m2l}{An ms2lib object.}

\item{pidx}{A pattern id or a frag_pattern object.}

\item{titles}{A vector of titles to be used. A default tile include id
and precursor will bne used by default}

\item{byPage}{The maximum number of spectra to be plotted by page.}

\item{...}{supplementary function to be passed to theplot function.}
}
\description{
PLotting the occurences of a fragPattern object inside
an ms2Lib object.
}
