% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexPlot.R
\name{plotPatterns}
\alias{plotPatterns}
\title{Plot patterns and occurences of an ms2Lib object}
\usage{
plotPatterns(m2l, ids = NULL, occurences = TRUE, full = FALSE,
  byPage = 9, ...)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{ids}{THe ids to be plotted or NULL if all the ids are supposed to be plot}

\item{occurences}{include the occurences}

\item{full}{Shall the full patterns be plotted (it can take some times)}

\item{byPage}{Maximum number of occurences by page.}
}
\description{
Plot patterns graph and occurences of an ms2Lib object.
}
