% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/complexPlot.R
\name{plotPatterns}
\alias{plotPatterns}
\alias{plotPatterns,ms2Lib-method}
\title{Plot patterns and spectra occurrences of an ms2Lib object}
\usage{
plotPatterns(
  m2l,
  ids = NULL,
  components = NULL,
  occurences = TRUE,
  full = FALSE,
  byPage = 9,
  titles = NULL,
  v_ggplot = TRUE,
  export_pdf = FALSE,
  path_inchi = NULL,
  infos_col = NULL,
  save_dir = "",
  ...
)

\S4method{plotPatterns}{ms2Lib}(
  m2l,
  ids = NULL,
  components = NULL,
  occurences = TRUE,
  full = FALSE,
  byPage = 9,
  titles = NULL,
  v_ggplot = TRUE,
  export_pdf = FALSE,
  path_inchi = NULL,
  infos_col = NULL,
  save_dir = ""
)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{ids}{The ids to be plotted or NULL to plot all ids}

\item{components}{An id giving the component of each spectrum to be plotted in first page}

\item{occurences}{Shall the occurrences be plotted as well as the pattern?}

\item{full}{Shall the full patterns be plotted?}

\item{byPage}{Maximum number of occurrences by page}

\item{titles}{A vector giving the titles of the MS/MS spectra}

\item{v_ggplot}{(default TRUE) if TRUE ggplot version for the plotting of the spectra}

\item{export_pdf}{(for ggplot version) if TRUE export the spectra to PDF}

\item{path_inchi}{name of a tabular file containing the inchi keys of the molecules in a column named "Name"; if this table is available, the 2D structures corresponding to the spectra will be retrieved from ChemSpider (webchem package) and displayed in the plot along the spectra (default NULL)}

\item{infos_col}{(for ggplot version) columns names from the spectra information to print}

\item{save_dir}{name of the directory to store the pdf files}

\item{...}{supplementary argument passed to the methods}
}
\value{
Nothing
}
\description{
`plotPatterns` plot the pattern graphs and supports (spectra occurrences) of an ms2Lib object.
}
\examples{
#Loading the data
data(m2l)

#plotting the patterns
plotPatterns(m2l,c("P30","P51"))
}
