% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{recognisedFormat}
\alias{recognisedFormat}
\title{Get the supported input format for mineMS2
For now, the only supported format is mgf.}
\usage{
recognisedFormat()
}
\value{
A character vector listing the supported format
}
\description{
Get the supported input format for mineMS2
For now, the only supported format is mgf.
}
\examples{
recognisedFormat()
}
