% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Selection function}
\usage{
select(m2l, ids, vals = c("P", "S", "L", "F"), ...)
}
\arguments{
\item{m2l}{An ms2Lib object.}

\item{ids}{Valid IDs of objects.}
}
\value{
The objects of the correspodning type.
}
\description{
Selection function
}
\examples{
print("Examples to be put here")
}
