% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\docType{methods}
\name{[,ms2Lib,ANY,ANY,ANY-method}
\alias{[,ms2Lib,ANY,ANY,ANY-method}
\title{Return the dag correspodning to a spectra or the modif.}
\usage{
\S4method{[}{ms2Lib,ANY,ANY,ANY}(x, i, j = NULL, ..., drop = TRUE)
}
\arguments{
\item{x}{An ms2Lib oject.}

\item{i}{The index, a string starting by S if it a spectrum, P if it's a pattern D if it's a dag, L if it's a loss
F if it's a fragment.}

\item{j}{unused.}

\item{...}{unused.}

\item{drop}{unused.}
}
\value{
A list containg the object.
}
\description{
Indexing function.
}
\examples{
print("Example to be put here")
}
