% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\name{vrange,ms2Lib-method}
\alias{vrange,ms2Lib-method}
\title{Calculate the range of iterations}
\usage{
\S4method{vrange}{ms2Lib}(m2l, type = c("S", "L", "P"), reduced = TRUE, as.number = FALSE)
}
\arguments{
\item{m2l}{An ms2Lib object}

\item{type}{"S" for spectra,"L" for m/z differences or "P" for patterns}

\item{reduced}{Used only if "type" is set to "P", shall the filtered pattern set be returned.}

\item{as.number}{If as.number is selected integer are returned without the prefix.}
}
\value{
A character vector giving the existing ids in the ms2Lib object,
}
\description{
Return the full range of iterations for different objects for an MS2lib object.
}
\examples{
#Loading the data
data(m2l)

#Range of iterations for spectra
vrange(m2l,"S")
}
