% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Lib.R
\docType{methods}
\name{vrange,ms2Lib-method}
\alias{vrange,ms2Lib-method}
\title{Return the range of iteration on an ms2Lib object.}
\usage{
\S4method{vrange}{ms2Lib}(m2l, type = c("S", "L", "P", "F"),
  reduced = TRUE)
}
\arguments{
\item{m2l}{AN ms2Lib object}

\item{type}{"S","L","P" or "F"}

\item{reduced}{Used only if "type" is set to "P".}
}
\value{
A character vector giving the existing ids.
}
\description{
Return the range of iteration on an ms2Lib object.
}
\examples{
print("Examples to be put here")
}
